% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNet-source.R
\name{aggregateCVSingle}
\alias{aggregateCVSingle}
\title{Saving cross-validation result as the cross-validation table into the
working directory and provide recommendation on the penalty term selection.}
\usage{
aggregateCVSingle(CVDir, SCCAmethod, K = 5, NumSubsamp = 500)
}
\arguments{
\item{CVDir}{A directory where the result is stored.}

\item{SCCAmethod}{The canonical correlation analysis method that is used in
the model.}

\item{K}{number of folds for cross-validation.}

\item{NumSubsamp}{The numebr of subsampling used.}
}
\description{
Saving cross-validation result as the cross-validation table into the
working directory
}
