% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{geneExpr}
\alias{geneExpr}
\title{A synthetic mRNA expression dataset.}
\format{An object of class \code{matrix} with 358 rows and 500 columns.}
\usage{
geneExpr
}
\description{
A matrix containing simulated mRNA expression levels for 358 subjects (rows)
and 500 features (columns).
}
\keyword{datasets}
