% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSmCCNet.R
\name{getCanCorMulti}
\alias{getCanCorMulti}
\title{Canonical Correlation Value for SmCCA}
\usage{
getCanCorMulti(X, CCcoef, CCWeight, Y)
}
\arguments{
\item{X}{list. A list of data each with n subjects.}

\item{CCcoef}{vector. A vector indicating weights for each pairwise canonical correlation.}

\item{CCWeight}{list. A list consists of canonical weight vectors corresponds to 
each element of X.}

\item{Y}{matrix. A matrix consists of phenotype.}
}
\description{
Get canonical correlation value for SmCCA given canonical weight vectors and scaling factor
}
\examples{
# getCanCorMulti()
}
