% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNet-source.R
\name{getCCAout_single}
\alias{getCCAout_single}
\title{Internal functions called by getRobustPseudoWeights_single.}
\usage{
getCCAout_single(X1, Trait, Lambda1, trace = FALSE)
}
\arguments{
\item{X1}{data.}

\item{Trait}{phenotype.}

\item{Lambda1}{penalty term}

\item{trace}{Whether to display CCA algorithm trace.}
}
\description{
Internal functions called by getRobustPseudoWeights_single.
}
\keyword{internal}
