% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicsSmCCA.R
\name{getRobustWeightsMulti}
\alias{getRobustWeightsMulti}
\title{Calculate the canonical  weights for SmCCA.}
\usage{
getRobustWeightsMulti(
  X,
  Trait,
  Lambda,
  s = NULL,
  NoTrait = FALSE,
  FilterByTrait = FALSE,
  SubsamplingNum = 1000,
  CCcoef = NULL,
  trace = FALSE,
  TraitWeight = FALSE
)
}
\arguments{
\item{X}{A list of omics data matrix with same set and order of subjects.}

\item{Trait}{An \eqn{n\times 1} trait data matrix for the same n subjects.}

\item{Lambda}{A vector of LASSO penalty parameters for \code{X}. \code{Lambda} needs
to be between 0 and 1.}

\item{s}{A vector corresponds to the number of omics data, specifying for each omics data, what is the 
percentage of omics feature being subsampled at each iteration.}

\item{NoTrait}{Logical, default is \code{FALSE}. Whether trait information
is provided.}

\item{FilterByTrait}{Logical, default is \code{FALSE}. Whether only the top
(\eqn{80\%}) features with highest correlation to the trait will be assigned
nonzero weights. The choice of \eqn{80\%} is based on the PMA package.}

\item{SubsamplingNum}{Number of feature subsamples. Default is 1000. Larger
number leads to more accurate results, but at a higher cost.}

\item{CCcoef}{Optional coefficients for the SmCCA pairwise canonical
correlations. If \code{CCcoef = NULL} (default), then the objective function
is the total sum of all pairwise canonical correlations. It can also be a
coefficient vector that follows the column order of \code{combn(K, 2)}.}

\item{trace}{Logical. Whether to display the CCA algorithm trace.}

\item{TraitWeight}{Whether to output the canonical weight for phenotype}
}
\value{
A canonical correlation weight matrix with \eqn{p_1+p_2} rows. Each
column is the canonical correlation weights based on subsampled \code{X1}
and \code{X2} features. The number of columns is \code{SubsamplingNum}.
}
\description{
Integrate multiple (can be more than 2) omics data type (and a quantitative phenotype), and calculate
the absolute canonical correlation weights for the omics features using
SmCCA. SmCCA takes into account a phenotype/trait.
SmCCA maximizes the total (weighted or unweighted) pairwise canonical
correlation weights between two omics data types and the trait. It requires
the trait to be quantitative. This function is the generalization of function
getRobustPseudoWeights().
}
\examples{


## For illustration, we only subsample 5 times.
set.seed(123)
X1 <- matrix(rnorm(600,0,1), nrow = 60)
X2 <- matrix(rnorm(600,0,1), nrow = 60)
Y <- matrix(rnorm(60,0,1), nrow = 60)
# Unweighted SmCCA
result <- getRobustWeightsMulti(X = list(X1, X2), Trait = Y, NoTrait = FALSE,
Lambda = c(0.5, 0.5),s = c(0.7, 0.7), SubsamplingNum = 20)
  
}
