% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNet-source.R
\name{getRobustWeightsMultiBinary}
\alias{getRobustWeightsMultiBinary}
\title{Run Sparse multiple Canonical Correlation Analysis and Obtain Canonical Weights (with Subsampling)}
\usage{
getRobustWeightsMultiBinary(
  X,
  Y,
  Between_Discriminate_Ratio = c(1, 1),
  SubsamplingPercent = NULL,
  CCcoef = NULL,
  LambdaBetween,
  LambdaPheno = NULL,
  SubsamplingNum = 1000,
  ncomp_pls = 3,
  EvalClassifier = FALSE,
  testData = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{A list of omics data each with n subjects.}

\item{Y}{A vector of binary variable, user needs to set the level of this 
variable to 0 and 1.}

\item{Between_Discriminate_Ratio}{A vector with length 2 specifying the relative importance
of between-omics relationship and omics-phenotype relationship.}

\item{SubsamplingPercent}{A vector with length equal to the number of omics data (\code{X}).}

\item{CCcoef}{A vector of scaling factors only for between-omics relationship.}

\item{LambdaBetween}{A vector of sparsity penalty value for each omics data.}

\item{LambdaPheno}{A penalty term when running the sparse PLS with phenotype.}

\item{SubsamplingNum}{Number of feature subsamples.}

\item{ncomp_pls}{Number of latent components for PLS.}

\item{EvalClassifier}{If TRUE, return latent factors for classification.}

\item{testData}{A list of testing omics data matrix.}

\item{verbose}{Logical; if TRUE, print progress/error messages, otherwise run silently.}
}
\value{
Canonical weight matrix or latent projections depending on EvalClassifier.
}
\description{
SmCCNet algorithm with multi-omics data and binary phenotype. This is a stepwise approach 
(1) use SmCCA to identify relationship between omics (exlude phenotype), (2) within highly connected omics features
selected in step 1, identify relationship between these selected omics features and phenotype of interest with 
sparse PLS. First, it computes PLSDA by assuming outcome is continuous to extract multiple latent factors, then uses latent factors to fit logistic regression, and weight latent factor by
regression parameters. Refer to multi-omics vignette for more detail.
}
