% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNet-source.R
\name{getRobustWeightsSingleBinary}
\alias{getRobustWeightsSingleBinary}
\title{One omics SmCCNet when the outcome variable is categorical}
\usage{
getRobustWeightsSingleBinary(
  X1,
  Trait,
  Lambda1,
  s1 = 0.7,
  SubsamplingNum = 1000,
  K = 3
)
}
\arguments{
\item{X1}{An \eqn{n\times p_1} data matrix (e.g. mRNA) with \eqn{p_1}
features and \eqn{n} subjects.}

\item{Trait}{An \eqn{n\times 1} trait data matrix for the same n subjects.}

\item{Lambda1}{LASSO penalty parameter for \code{X1}. \code{Lambda1} needs
to be between 0 and 1.}

\item{s1}{Proportion of mRNA features to be included, default at \code{s1 =
0.7}. \code{s1} needs to be between 0 and 1.}

\item{SubsamplingNum}{The numebr of subsampling used.}

\item{K}{Number of hidden components for PLSDA}
}
\description{
One omics SmCCNet when there is only one omics, and the outcome variable is
categorical, implement SPLS instead of the regular canonical correlation
analysis.
}
\examples{


X <- matrix(rnorm(600,0,1), nrow = 60)
Y <- rbinom(60,1,0.5)
Ws <- getRobustWeightsSingleBinary(X1 = X, Trait = as.matrix(Y), Lambda1 = 0.8, 
0.7, SubsamplingNum = 10)

}
