% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNetSource.R
\name{plotMultiOmicsNetwork}
\alias{plotMultiOmicsNetwork}
\title{Plot multi-omics modules based on similarity matrix derived from pseudo
canonical weights and pairwise feature correlations.}
\usage{
plotMultiOmicsNetwork(Abar, CorrMatrix, multiOmicsModule, ModuleIdx, P1,
  EdgeCut, FeatureLabel = NULL, AddCorrSign = TRUE, SaveFile = NULL,
  ShowType1Label = TRUE, ShowType2Label = TRUE, PlotTitle = "",
  NetLayout = "lgl", ShowNodes = TRUE, VertexLabelCex = 1,
  VertexSize = 1)
}
\arguments{
\item{Abar}{A p\eqn{\times}p similary matrix for both omics data types
based on pseudo canonical weights. All entries are non-negative.}

\item{CorrMatrix}{A p\eqn{\times}p correlation matrix that provides sign
information for the network.}

\item{multiOmicsModule}{A list of multi-omics modules.}

\item{ModuleIdx}{Index for the module to be plotted. It can not exceed the
length of multiOmicsModule.}

\item{P1}{Total number of features for the first omics data type.}

\item{EdgeCut}{A numerical value between 0 and 1, indicating an edge
threshold for the network. Any features (network nodes) without any edge
strength that passes the threshold are excluded from the figure.}

\item{FeatureLabel}{A 1\eqn{\times}p vector indicating feature names. If
FeatureLabel = NULL (default), the feature names will be indices 1
through p, where p is the total number of omics features.}

\item{AddCorrSign}{Logical. Whether to add a positive or negative sign to
each network edge based on pairwise feature correlations.}

\item{SaveFile}{A pdf file name for the figure output. If SaveFile = NULL
(default), the figure will not be saved.}

\item{ShowType1Label}{Logical. Whether to label the network nodes for the
first omics data type.}

\item{ShowType2Label}{Logical. Whether to label the network nodes for the
second omics data type.}

\item{PlotTitle}{A title for the figure. Default is without any title.}

\item{NetLayout}{Graphical layout for the network. Possible options are
"circle", "sphere" for 3D sphere, "fr" for Fruchterman-Reinhold, and "lgl"
for the LGL algorithm. Refer to igraph manual for more details on the
layout options.}

\item{ShowNodes}{Logical. Whether to show network nodes.}

\item{VertexLabelCex}{Scaling factor for the vertex labels.}

\item{VertexSize}{Size of the vertices.}
}
\value{
A multi-omics network figure.
}
\description{
Plot multi-omics modules based on similarity matrix derived from pseudo
canonical weights and pairwise feature correlations.
}
\examples{
set.seed(123)
w <- rnorm(5)
w <- w/sqrt(sum(w^2))
abar <- getAbar(w)
modules <- getMultiOmicsModules(abar, P1 = 2, CutHeight = 0.5)
x <- cbind(X1[ ,1:2], X2[ , 1:3])
corr <- cor(x)

plotMultiOmicsNetwork(abar, corr, modules, ModuleIdx = 1, P1 = 2,
  EdgeCut = 0)

}
