% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSmCCNet.R
\name{scalingFactorInput}
\alias{scalingFactorInput}
\title{Scaling Factor Input}
\usage{
scalingFactorInput(DataType = NULL)
}
\arguments{
\item{DataType}{A vector that contains the the data type of each omics data.}
}
\description{
Input the vector of the type of dataset, and return prompts that ask the user to supply the scaling 
factor intended for SmCCNet algorithm to prioritize the correlation structure of 
interest. All scaling factor values supplied should be numeric and nonnegative.
}
\examples{
# not run
# scalingFactorInput(c('gene','mirna', 'phenotype'))

}
