% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNet-source.R
\name{summarizeNetSHy}
\alias{summarizeNetSHy}
\title{NetSHy Summarization Score}
\usage{
summarizeNetSHy(X, A, is_alpha = TRUE, npc = 1)
}
\arguments{
\item{X}{Data matrix of size N by P, where N is the number of subjects and P is the number 
of features.}

\item{A}{Corresponding adjacency matrix of size P by P.}

\item{is_alpha}{Whether TOM matrix is considered as part of the network topology, 
default is set to TRUE}

\item{npc}{Number of principal components used to summarize the network.}
}
\value{
a list consists of [[1]] Subject-level network summarization score,
[[2]] Principal component importance information such as percent of variance explained,
(3) Principal component feature-level loadings.
}
\description{
Implement NetSHy network summarization via a hybrid approach to 
summarize network by considering the network topology with laplacian matrix (and TOM matrix).
}
\examples{
# simulate omics data
OmicsData <- matrix(rnorm(200,0,1), nrow = 10, ncol = 20)
# simulate omics adjacency matrix
set.seed(123)
w <- rnorm(20)
w <- w/sqrt(sum(w^2))
abar <- getAbar(w, P1 = 2, FeatureLabel = NULL)
# extract NetSHy summarization score
netshy_score <- summarizeNetSHy(OmicsData, abar)

}
\references{
Vu, T., Litkowski, E. M., Liu, W., Pratte, K. A., Lange, L., Bowler, R. P., ... & Kechris, K. J. (2023). NetSHy: network summarization via a hybrid approach leveraging topological properties. Bioinformatics, 39(1), btac818.
}
