% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{X1}
\alias{X1}
\title{Two simulated omics datasets and a quantitative phenotype for the same 358 subjects.}
\format{A mRNA matrix with 358 rows and 500 columns}
\usage{
X1
}
\description{
Two simulated omics datasets and a quantitative phenotype for the same 358 subjects.
}
\keyword{datasets}
