% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{X2}
\alias{X2}
\title{A synthetic miRNA expression dataset.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 358 rows and 100 columns.
}
\usage{
X2
}
\description{
A matrix containing simulated miRNA expression levels for 358 subjects (rows)
and 100 features (columns).
}
\keyword{datasets}
