% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Y}
\alias{Y}
\title{A phenotype matrix with 358 rows and 1 column}
\format{An object of class \code{matrix} with 358 rows and 1 columns.}
\usage{
Y
}
\description{
A phenotype matrix with 358 rows and 1 column
}
\keyword{datasets}
