% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNetSource.R
\name{getAbar}
\alias{getAbar}
\title{Compute the similarity matrix based on the outer products of absolute canonical
weights.}
\usage{
getAbar(Ws, FeatureLabel = NULL)
}
\arguments{
\item{Ws}{A canonical weight vector or matrix. If Ws is a matrix, then each
column corresponds to one weight vector.}

\item{FeatureLabel}{A 1\eqn{\times}p vector indicating feature names. If
FeatureLabel = NULL (default), the feature names will be indices 1
through p, where p is the total number of omics features.}
}
\value{
A p\eqn{\times}p symmetric non-negative matrix.
}
\description{
Compute the similarity matrix based on the outer products of absolute canonical
weights.
}
\examples{
w <- matrix(rnorm(6), nrow = 3)
Ws <- apply(w, 2, function(x)return(x/sqrt(sum(x^2))))
abar <- getAbar(Ws)

}
