% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicsSmCCA.R
\name{getCanWeightsMulti}
\alias{getCanWeightsMulti}
\title{Get Canonical Weight SmCCA Algorithm (No Subsampling)}
\usage{
getCanWeightsMulti(
  X,
  Trait = NULL,
  Lambda,
  CCcoef = NULL,
  NoTrait = TRUE,
  trace = FALSE,
  TraitWeight = FALSE
)
}
\arguments{
\item{X}{A list of omics data each with n subjects.}

\item{Trait}{An n by k trait data for the same samples (k >=1).}

\item{Lambda}{A vector of LASSO pentalty parameters, need to be between 0 and 1,
should also have the same length as the length of X.}

\item{CCcoef}{A vector indicating weights for each pairwise canonical correlation.}

\item{NoTrait}{Logical. Whether trait information is provided to Trait will be assigned nonzero weights.}

\item{trace}{Logical. Whether to display CCA algorithm trace.}

\item{TraitWeight}{Logical. Whether to include canonical weight for trait, default is set to FALSE}
}
\description{
Run Sparse Multiple Canonical Correlation Analysis (SmCCA) and return
canonical weight.
}
\examples{
# not run
# X <- list(X1,X2)
# result <- getCanWeightsMulti(X, Trait = as.matrix(Y), Lambda = c(0.5,0.5), NoTrait = FALSE)
# result <- getCanWeightsMulti(X, Trait = NULL, Lambda = c(0.5,0.5), NoTrait = TRUE)
}
