% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmCCNetSource.R
\name{getRobustPseudoWeights}
\alias{getRobustPseudoWeights}
\title{Compute canonical weights based on sparse multiple canonical correlations
(SmCCA), sparse supervised canonical correlations (SsCCA), or sparse
canonical correlations (SCCA).}
\usage{
getRobustPseudoWeights(X1, X2, Trait, Lambda1, Lambda2, s1 = 0.9, s2 = 1,
  NoTrait = FALSE, FilterByTrait = FALSE, Bipartite = FALSE,
  SubsamplingNum = 100, PartitionNum = 100, CCcoef = NULL,
  trace = FALSE)
}
\arguments{
\item{X1}{An n\eqn{\times}p1 data matrix (e.g. mRNA) with p1 features and n
subjects.}

\item{X2}{An n\eqn{\times}p2 data matrix (e.g. miRNA) with p2 features and n
subjects.}

\item{Trait}{An n\eqn{\times}1 trait data matrix for the same n subjects.}

\item{Lambda1}{LASSO pentalty parameter for X1, need to be between 0 and 1.}

\item{Lambda2}{LASSO pentalty parameter for X2, need to be between 0 and 1.}

\item{s1}{Proportion of mRNA features to be included.}

\item{s2}{Proportion of miRNA features to be included.}

\item{NoTrait}{Logical. Whether trait information is provided.}

\item{FilterByTrait}{Logical. Whether only the top 80% features with highest
correlation to the trait will be assigned nonzero weights.}

\item{Bipartite}{Logical. Whether to include random partition.}

\item{SubsamplingNum}{Number of feature subsamples. Larger number leads to
more accurate results, but at a higher cost. We recommend to subsample at
least 1000 times.}

\item{CCcoef}{Optional coefficients for the pairwise canonical correlations
(CC). If CCcoef = NULL (default), then the objective function is the total
sum of all pairwise CC. It can also be a coefficient vector that follows
the column order of combn(K, 2).}

\item{trace}{Logical. Whether to display CCA algorithm trace.}

\item{PartitionNum:}{Number of random partitions for each . This is only used if
Bipartite = FALSE.}
}
\value{
A canonical weight matrix with p1+p2 rows. Each column is the
  canonical weights based on subsampled X1 and X2 features.The column number
  equals to SubsamplineNum.
}
\description{
SmCCA and SsCCA take into account of a phenotype/trait. SmCCA maximizes the
total (weighted or unweighted) pairwise canonical weights between two omics
data type and the trait. It requires the trait to be quantitative. SsCCA
prioritizes omics features based on the trait, and assigns non-zero canonical
weights to features that are more correlated to the trait. SCCA does not use
any trait information for computing the canonical weights. All of these three
methods are included in the function, along with an omic feature subsampling
scheme.
}
\examples{
\donttest{
## For illustration, we only subsample 5 times.
set.seed(123)

# SmCCA
W1 <- getRobustPseudoWeights(X1, X2, Trait = Y, Lambda1 = 0.05,
  Lambda2 = 0.05, s1 = 0.7, s2 = 0.9, NoTrait = FALSE, FilterByTrait = FALSE,
  Bipartite = FALSE, SubsamplingNum = 100, PartitionNum = 5,
  CCcoef = NULL, trace = FALSE)

# SsCCA
W2 <- getRobustPseudoWeights(X1, X2, Trait = Y, Lambda1 = .05, Lambda2 = 0.5
  s1 = 0.7, s2 = 0.9, NoTrait = FALSE, FilterByTrait = TRUE,
  Bipartite = FALSE, SubsamplingNum = 100, PartitionNum = 5,
  CCcoef = NULL, trace = FALSE)

# SCCA
W3 <- getRobustPseudoWeights(X1, X2, Trait = Y, Lambda1 = 0.05,
  Lambda2 = 0.05, s1 = 0.7, s2 = 0.9, NoTrait = TRUE, Bipartite = FALSE,
  SubsamplingNum = 100, PartitionNum = 5, CCcoef = NULL, trace = FALSE)
  }

}
