% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{X1}
\alias{X1}
\title{A synthetic mRNA expression dataset.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 358 rows and 500 columns.
}
\usage{
X1
}
\description{
A matrix containing simulated mRNA expression levels for 358 subjects (rows)
and 500 features (columns).
}
\keyword{datasets}
