% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{X2}
\alias{X2}
\title{A miRNA matrix with 358 rows and 100 columns}
\format{An object of class \code{matrix} with 358 rows and 100 columns.}
\usage{
X2
}
\description{
A miRNA matrix with 358 rows and 100 columns
}
\keyword{datasets}
