% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Y}
\alias{Y}
\title{A synthetic phenotype dataset.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 358 rows and 1 columns.
}
\usage{
Y
}
\description{
A matrix containing simulated quantitative phenotype measures for 358 
subjects (rows).
}
\keyword{datasets}
